donut/requirements_donut.txt
torch
transformers
Pillow
sentencepiece
python-dotenv